+++ apps/portfolio/src/app/_providers/query-client-provider.tsx
@@ -2,8 +2,6 @@
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
-import type React from 'react'
-
 const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
