+++ apps/portfolio/src/app/layout.tsx
@@ -1,12 +1,18 @@
 import './_styles/global.css'
 
+import { Suspense } from 'react'
+
 import { ToastContainer } from '@status-im/components'
 import { Analytics } from '@vercel/analytics/next'
 import { Inter } from 'next/font/google'
 
 import { Navbar } from './_components/navbar'
 import { NotAllowed } from './_components/not-allowed'
-import { Providers } from './_providers'
+import { AccountsProvider } from './_providers/accounts-context'
+import { ConnectKitProvider } from './_providers/connectkit-provider'
+import { QueryClientProvider } from './_providers/query-client-provider'
+import { StatusProvider } from './_providers/status-provider'
+import { WagmiProvider } from './_providers/wagmi-provider'
 
 import type { Metadata } from 'next'
 
@@ -35,20 +41,28 @@ export default async function PortfolioLayout(props: Props) {
       </head>
       <body data-customisation="blue" suppressHydrationWarning>
         <div id="app" className="isolate">
-          <Providers>
-            <>
-              <Navbar />
-              <div className="px-1">
-                <div className="hidden flex-1 flex-col 2md:flex xl:pb-1">
-                  <div className="flex h-[calc(100vh-60px)] flex-col overflow-clip rounded-[24px] bg-white-100">
-                    {children}
-                  </div>
-                </div>
-                <NotAllowed />
-                <ToastContainer />
-              </div>
-            </>
-          </Providers>
+          <StatusProvider>
+            <WagmiProvider>
+              <QueryClientProvider>
+                <Suspense fallback={<div>Loading...</div>}>
+                  <AccountsProvider>
+                    <ConnectKitProvider>
+                      <Navbar />
+                      <div className="px-1">
+                        <div className="hidden flex-1 flex-col 2md:flex xl:pb-1">
+                          <div className="flex h-[calc(100vh-60px)] flex-col overflow-clip rounded-[24px] bg-white-100">
+                            {children}
+                          </div>
+                        </div>
+                        <NotAllowed />
+                        <ToastContainer />
+                      </div>
+                    </ConnectKitProvider>
+                  </AccountsProvider>
+                </Suspense>
+              </QueryClientProvider>
+            </WagmiProvider>
+          </StatusProvider>
         </div>
         <Analytics debug={false} />
         {/* <VercelToolbar /> */}
