+++ apps/wallet/src/main.tsx
@@ -5,6 +5,7 @@ import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
 
 import { APIClientProvider } from './providers/api-client'
+import { QueryClientProvider } from './providers/query-client'
 import { RouterProvider } from './providers/router'
 import { StatusProvider } from './providers/status'
 
@@ -21,11 +22,13 @@ if (!rootElement.innerHTML) {
   const root = ReactDOM.createRoot(rootElement)
   root.render(
     <StrictMode>
-      <StatusProvider>
-        <APIClientProvider>
-          <RouterProvider />
-        </APIClientProvider>
-      </StatusProvider>
+      <QueryClientProvider>
+        <StatusProvider>
+          <APIClientProvider>
+            <RouterProvider />
+          </APIClientProvider>
+        </StatusProvider>
+      </QueryClientProvider>
     </StrictMode>,
   )
 }
