+++ apps/wallet/src/routes/__root.tsx
@@ -1,13 +1,8 @@
 // import { Suspense } from 'react'
 
 // import OnboardingPage from '../../../portfolio/src/app/page'
-import { Button, ToastContainer } from '@status-im/components'
+import { ToastContainer } from '@status-im/components'
 import { Navbar } from '@status-im/wallet/components'
-// import { TanStackRouterDevtools } from '@tanstack/router-devtools'
-// import { NotAllowed } from '../../../portfolio/src/app/_components/not-allowed'
-// import { AccountsProvider } from '../../../portfolio/src/app/_providers/accounts-context'
-// import { ConnectKitProvider } from '../../../portfolio/src/app/_providers/connectkit-provider'
-import { QueryClientProvider, WagmiProvider } from '@status-im/wallet/providers'
 // import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import {
   createRootRouteWithContext,
@@ -17,7 +12,13 @@ import {
   redirect,
 } from '@tanstack/react-router'
 
+// import { TanStackRouterDevtools } from '@tanstack/router-devtools'
+// import { NotAllowed } from '../../../portfolio/src/app/_components/not-allowed'
+// import { AccountsProvider } from '../../../portfolio/src/app/_providers/accounts-context'
+// import { ConnectKitProvider } from '../../../portfolio/src/app/_providers/connectkit-provider'
+import { QueryClientProvider } from '../../../portfolio/src/app/_providers/query-client-provider'
 // import { StatusProvider } from '../../../portfolio/src/app/_providers/status-provider'
+import { WagmiProvider } from '../../../portfolio/src/app/_providers/wagmi-provider'
 import { Link } from '../components/link'
 import { apiClient } from '../providers/api-client'
 import { PendingTransactionsProvider } from '../providers/pending-transactions-context'
@@ -155,11 +156,7 @@ function RootComponent() {
             <WalletProvider>
               <PendingTransactionsProvider>
                 <div className="flex min-h-[56px] items-center px-2">
-                  <Navbar
-                    hasFeedback
-                    linkComponent={Link}
-                    rightSlot={<Button variant="outline">Test</Button>}
-                  />
+                  <Navbar hasFeedback linkComponent={Link} />
                 </div>
                 <div className="px-1">
                   <div className="flex-1 flex-col 2md:flex xl:pb-1">
