+++ packages/wallet/package.json
@@ -41,11 +41,6 @@
       "types": "./dist/src/utils/index.d.ts",
       "import": "./dist/utils/index.es.js",
       "require": "./dist/utils/index.cjs.js"
-    },
-    "./providers": {
-      "types": "./dist/src/providers/index.d.ts",
-      "import": "./dist/providers/index.es.js",
-      "require": "./dist/providers/index.cjs.js"
     }
   },
   "files": [
@@ -69,10 +64,7 @@
     "prepack": "pnpm build"
   },
   "peerDependencies": {
-    "@tanstack/react-query": "5.66.0",
-    "react": "^19.1.0",
-    "viem": "^2.21.1",
-    "wagmi": "2.15.2"
+    "react": "^18.2.0"
   },
   "dependencies": {
     "@hookform/resolvers": "^3.3.4",
@@ -140,8 +132,8 @@
     "globals": "^15.12.0",
     "postcss": "^8.4.33",
     "prettier-plugin-tailwindcss": "^0.6.6",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "rollup-plugin-preserve-directives": "^0.4.0",
     "tailwind-scrollbar-utilities": "^0.2.0",
     "tailwindcss": "^3.4.11",
