--- packages/wallet/src/providers/index.tsx
@@ -1,2 +1,0 @@
-export { QueryClientProvider } from './query-client-provider'
-export { wagmiConfig, WagmiProvider } from './wagmi-provider'
