+++ packages/wallet/vite.config.ts
@@ -22,7 +22,6 @@ export default defineConfig(({ mode }) => {
           'hooks/index': './src/hooks/index.ts',
           'constants/index': './src/constants/index.ts',
           'utils/index': './src/utils/index.ts',
-          'providers/index': './src/providers/index.tsx',
         },
         formats: ['es', 'cjs'],
         fileName: format => {
